Add files via upload
Auto-capture Selfie by Detecting Smile:

This feature allows you to automatically capture a selfie when the camera detects a smile. This is a great way to take quick and easy selfies without having to worry about pressing the shutter button at the right time.

To use this feature, simply open the camera app and enable the "Auto-capture Selfie" setting. Then, hold the camera up to your face and smile. The camera will automatically take a photo when it detects a smile.

This feature is perfect for people who want to take quick and easy selfies, but don't want to worry about pressing the shutter button at the right time. It is also a great way to take selfies with friends or family, as you can all just smile and the camera will take a photo of you all together.